+++ crates/base/core/src/text/error.rs
@@ -32,7 +32,7 @@ define_error! { individual:
         pub error_len: Option<usize>
     }
     +location: "text", +tag: _tags!(text),
-    DOC_INVALID_UTF8 = "Invalid Utf-8 found while interpreting a byte sequence.\n\n
+    DOC_INVALID_UTF8 = "Invalid UTF-8 found while interpreting a byte sequence.\n\n
 This is basically a replication of `core::str::`[`Utf8Error`]`.",
     self+f => if let Some(len) = self.error_len {
         write!(f, "Invalid UTF-8, valid up to: {}, len: {len}", self.valid_up_to)