+++ src/text/parse/mod.rs
@@ -9,6 +9,9 @@ crate::structural_mods! { // _reexports
     _reexports {
         pub use super::_reexport_core::*;
         #[doc(inline)]
-        pub use devela_base_core::text::ByteSearch;
+        pub use devela_base_core::text::{
+            ByteSearch,
+            TextParseError, TextParseErrorKind, TextScanner,
+        };
     }
 }